Move ButtonTooltip into shared/Button
