+++ frontend/src/components/shared/Button/index.tsx
@@ -18,4 +18,5 @@
 import { ButtonProps as ButtonPropsT } from "./Button"
 
 export { default, Kind, Size } from "./Button"
+export { ButtonTooltip } from "./ButtonTooltip"
 export type ButtonProps = ButtonPropsT
