+++ frontend/src/components/shared/Button/styled-components.ts
@@ -205,3 +205,17 @@ export const StyledIconButton = styled(StyledBaseButton)<RequiredButtonProps>(
     }
   }
 )
+
+export const StyledTooltipNormal = styled.div(({ theme }) => ({
+  display: "block",
+  [`@media (max-width: ${theme.breakpoints.sm})`]: {
+    display: "none",
+  },
+}))
+
+export const StyledTooltipMobile = styled.div(({ theme }) => ({
+  display: "none",
+  [`@media (max-width: ${theme.breakpoints.sm})`]: {
+    display: "block",
+  },
+}))
