+++ frontend/src/components/widgets/Button/Button.test.tsx
@@ -22,13 +22,13 @@ import { WidgetStateManager } from "src/lib/WidgetStateManager"
 import UIButton from "src/components/shared/Button"
 
 import { Button as ButtonProto } from "src/autogen/proto"
-import Button, { ButtonProps } from "./Button"
+import Button, { Props } from "./Button"
 
 jest.mock("src/lib/WidgetStateManager")
 
 const sendBackMsg = jest.fn()
 
-const getProps = (elementProps: Partial<ButtonProto> = {}): ButtonProps => ({
+const getProps = (elementProps: Partial<ButtonProto> = {}): Props => ({
   element: ButtonProto.create({
     id: "1",
     label: "Label",
