--- frontend/src/components/widgets/Button/styled-components.ts
@@ -1,15 +1,0 @@
-import styled from "@emotion/styled"
-
-export const StyledTooltipNormal = styled.div(({ theme }) => ({
-  display: "block",
-  [`@media (max-width: ${theme.breakpoints.sm})`]: {
-    display: "none",
-  },
-}))
-
-export const StyledTooltipMobile = styled.div(({ theme }) => ({
-  display: "none",
-  [`@media (max-width: ${theme.breakpoints.sm})`]: {
-    display: "block",
-  },
-}))
